Implemented a bunch of PatchBasisTests, which will right now fail with an exception, because Mesh doesn't yet have any PatchBasis support implemented.  (Doing test-driven development here.)
Replaced a 2-argument refine() method in Mesh with 1-argument pRefine(), which is all it's ever supported anyhow (and is good symmetry with Mesh::hRefine()).

Added Mesh::physicalCellNodesForCell(int cellID) method, which returns a FieldContainer that can be used for doing single-element computations, even for inactive elements (useful for testing PatchBasis).



git-svn-id: https://svn.ices.utexas.edu/repos/camellia@189 36274a68-8941-4c93-80ef-affaed5139da