Added primitive support for writeTOVTK.
The code is very ugly and only supports 2D quads at the moment. Ideally I
would like to link to the VTK library, which would streamline a lot of the
output process and allow output in a much more compact binary format. This
will be very important once we start using bigger meshes.