Many changes, largely in pursuit of errors in the L^2 projection.  Added a test to SolutionTests which fails: this projects the Kovasznay solution onto a constant mesh.  Also fixed an issue in which the projected solution would have multi-valued traces and fluxes.
Rewrote the Projector::project() method that takes a FunctionPtr but no IP to call the one that takes an IP.

It is worth noting that after my recent changes the nature of the Kovasznay failure has changed, a little: the values are still incorrect, but they are different (as it happens, closer, at least in the case I was paying attention to).

Switched DofOrdering to use a set<int> for its VarIDs.